refactor: Stadium API 변경
